requirements.txt
flask
flask-login
flask-wtf
boto3
gunicorn
psycopg2
pyyaml
